requirements.txt
cycler==0.10.0
Django==2.2.1
django-cors-headers==3.0.2
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
mlxtend==0.16.0
mysqlclient==1.4.2.post1
numpy==1.16.4
pandas==0.24.2
peewee==3.9.6
Pillow==6.0.0
pkg-resources==0.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.2
python-logstash==0.4.6
pytz==2019.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
SQLAlchemy==1.3.4
sqlparse==0.3.0
pytest
